requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
fastapi==0.112.2
pydantic==2.8.2
uvicorn[standard]==0.30.6
PyYAML==6.0.2
pytest==8.3.2
httpx==0.27.2
slowapi==0.1.9
python-multipart==0.0.9
cryptography==41.0.8
transformers>=4.20.0
torch>=2.0.0
accelerate==0.34.2
sentencepiece==0.2.0
kagglehub==0.3.2
strands-agents==1.4.0
